Create PrivateSecuritiesSettlementV4-2021-05-25-11pmEST.gs
Contains most up-to-date Processing & Settling functions. These functions no longer rely on pulling row numbers of customers from formulas in Today's Trades and instead use a function to figure out the correct row. Still needs to create a function that can figure out the column number of securities in MB (for now this relies on index formula in Todays Trades)